Implementation of Theiactl and policy recommendation CLI
In this commit, we add the implementation of Theiactl, the command-line
tool for Theia based on cobra framework.
Policy recommendation CLI is also included. It consists of 3 commands
which let users could start a new policy recommendation job, check the
current status of a previous job, and get the recommendation result of a
previous job respectively.

Signed-off-by: Yongming Ding <dyongming@vmware.com>